As mentioned in https://github.com/google/xi-editor/issues/456 and as I discussed with @cmyr on IRC, it would be better to not have the explanation only in the issues. The issues are sort of a stack and having this explanation only in a 'we need to refactor' issue hides it from any potential contributor.