@cmyr 2018-04-08
Awesome, thanks for following up on this. I am however wondering where this should live; in particular I'm worried about discoverability. I'm not sure anyone will know to look in `rope/docs`. This is almost-but-not-quite part of the rope science series in the top level docs, but I'm not sure.

Maybe my suggestion right now would be that we go with your plan here, but explicitly mention this file somewhere in `rope.rs`, so people who are curious have a better chance of finding it?