@mqzry 2018-04-08
I think this is a step in the right direction for discoverability from a contributors perspective.
Listing the contents of the root of the rope crate now will have a `docs` dir, which is the obvious place to look further if interested.

The fact that you have to look in rope/docs AND in the science series to get a clue what is happening in the rope crate is not  optimal.  Ideally the rope crate should have all documentation contained within for understanding it, with links to some resources(papers/specs/) that explain everything in great detail for the interested. These resources should not be globally linked to, but whenever some topic is used we should point to that specific section. So instead of 'For info about x, see science_series_url' on the top of the file,  it would be better to say 'For more info about atomic metrics, see science_series_url#metrics'.