@cmyr 2018-04-09
Okay, that sounds reasonable. I'm going to leave this up for @raphlinus because this is very much his domain, and he might have some other ideas, but lgtm.